+++ application/controllers/Welcome.php
@@ -29,5 +29,25 @@ class Welcome extends Application
 
 		$this->render();
 	}
+	
+	public function random()
+	{
+		$num = rand(0,4);
+		
+		// this is the view we want shown
+		$this->data['pagebody'] = 'homepage';
+
+		// build the list of authors, to pass on to our view
+		$source = $this->quotes->all();
+		$authors = array ();
+		foreach ($source as $record)
+		{
+			$authors[] = array ('who' => $record['who'], 'mug' => $record['mug'], 'href' => $record['where'], 'what' => $record['what']);
+		}
+		
+		$this -> data['authors'] = array($authors[rand(0,sizeof($authors)-1)]);
+		
+		$this->render();
+	}
 
 }